# crm-it-shkola-rtk/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workflow/templates": {
      "get": {
        "operationId": "WorkflowController_getTemplates",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowTemplateWithActiveVersionDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Список шаблонов workflow с активной версией (статусы+переходы)",
        "tags": [
          "workflow"
        ]
      },
      "post": {
        "operationId": "WorkflowController_createTemplate",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateVersionDto"
                }
              }
            }
          }
        },
        "summary": "Создать шаблон workflow вместе с первой версией (статусы+переходы)",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/templates/{id}": {
      "put": {
        "operationId": "WorkflowController_updateTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateVersionDto"
                }
              }
            }
          }
        },
        "summary": "Редактирование шаблона — НЕ мутирует текущую версию, а создаёт новую WorkflowTemplateVersion (isActive=true), предыдущую помечает isActive=false. Существующие InteractionInstance остаются привязаны к своей версии и не переезжают на новую автоматически.",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/templates/{id}/graph": {
      "put": {
        "description": "Для экрана-редактора (список/форма): не набор точечных CRUD-вызовов на статус/переход, а один граф. statuses/transitions ссылаются друг на друга по полю id внутри ЭТОГО запроса (client-side ключ узла, не id в БД) — так редактору не нужно сначала создавать статусы, чтобы узнать их серверные id, а потом создавать переходы. Как и PUT /workflow/templates/{id}: НЕ мутирует текущую версию, создаёт новую (isActive=true), предыдущую помечает isActive=false; существующие InteractionInstance остаются на своей версии. migrateInstances не поддерживается — используйте PUT /workflow/templates/{id} для переноса.",
        "operationId": "WorkflowController_updateTemplateGraph",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowGraphDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateVersionDto"
                }
              }
            }
          }
        },
        "summary": "Заменить граф шаблона целиком (все статусы и переходы одним запросом, в одной транзакции)",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/template-versions/{id}": {
      "get": {
        "operationId": "WorkflowController_getTemplateVersionById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0000000-0000-4000-8000-000000000002",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateVersionDto"
                }
              }
            }
          }
        },
        "summary": "Версия шаблона со статусами и переходами",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/instances": {
      "get": {
        "operationId": "WorkflowController_getInstances",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InteractionInstanceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Список взаимодействий с вузами (видимость по роли — как в catalogs)",
        "tags": [
          "workflow"
        ]
      },
      "post": {
        "operationId": "WorkflowController_createInstance",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInteractionInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionInstanceDto"
                }
              }
            }
          }
        },
        "summary": "Создать взаимодействие на активной версии активного шаблона для вуза",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/instances/{id}": {
      "get": {
        "operationId": "WorkflowController_getInstanceById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b3000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionInstanceDto"
                }
              }
            }
          }
        },
        "summary": "Взаимодействие по идентификатору",
        "tags": [
          "workflow"
        ]
      },
      "patch": {
        "description": "Все поля необязательны, но хотя бы одно обязательно. Статус здесь менять нельзя (400 WORKFLOW_STATUS_VIA_TRANSITION_ONLY) — только POST /workflow/instances/{id}/transition с проверкой графа переходов. Вуз — PATCH …/assignment. Ответственного меняют Руководитель (в своей команде) и Администратор; заметку и продукт — любая роль в своей зоне видимости.",
        "operationId": "WorkflowController_updateInstance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b3000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "rukovoditel"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInteractionInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionInstanceDto"
                }
              }
            }
          }
        },
        "summary": "Точечно обновить ответственного / ИТ-продукт / заметку взаимодействия (без перехода по статусам)",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/instances/{id}/assignment": {
      "patch": {
        "description": "Не мутирует статус/историю — только universityId/responsibleUserId (+пересчитывает needsReview). Хотя бы одно из полей обязательно. Если передан только universityId — ответственным по умолчанию становится кАМ этого вуза (University.kamId), как и при создании инстанса вручную. Руководитель может назначать вуз/ответственного только в пределах своей команды.",
        "operationId": "WorkflowController_patchAssignInstance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b3000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "rukovoditel"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignInteractionInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionInstanceDto"
                }
              }
            }
          }
        },
        "summary": "Назначить вуз и/или ответственного инстансу без них (needsReview=true из integrations/sync)",
        "tags": [
          "workflow"
        ]
      },
      "put": {
        "deprecated": true,
        "description": "Не мутирует статус/историю — только universityId/responsibleUserId (+пересчитывает needsReview). Хотя бы одно из полей обязательно. Если передан только universityId — ответственным по умолчанию становится кАМ этого вуза (University.kamId), как и при создании инстанса вручную. Руководитель может назначать вуз/ответственного только в пределах своей команды.",
        "operationId": "WorkflowController_assignInstance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b3000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "rukovoditel"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignInteractionInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionInstanceDto"
                }
              }
            }
          }
        },
        "summary": "Назначить вуз и/или ответственного инстансу без них (needsReview=true из integrations/sync)",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/instances/{id}/transition": {
      "post": {
        "operationId": "WorkflowController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b3000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionInteractionInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionInstanceDto"
                }
              }
            }
          }
        },
        "summary": "Переход в новый статус — проверяется по WorkflowTransition версии этого инстанса (400, если переход не разрешён), пишет append-only StatusHistoryEntry и обновляет currentStatusId. КАМ — только для своих вузов.",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/instances/{id}/history": {
      "get": {
        "operationId": "WorkflowController_getInstanceHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b3000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusHistoryEntryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Вся история переходов по статусам взаимодействия, по возрастанию времени",
        "tags": [
          "workflow"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM ИТ Школа РТК",
...
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "WorkflowTemplateVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b0000000-0000-4000-8000-000000000002"
          },
          "versionNumber": {
            "type": "number",
            "example": 2
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "workflowTemplateId": {
            "type": "string",
            "example": "b0000000-0000-4000-8000-000000000001"
          },
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStatusDto"
            }
          },
          "transitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowTransitionDto"
            }
          }
        },
        "required": [
          "id",
          "versionNumber",
          "isActive",
          "workflowTemplateId",
          "statuses",
          "transitions"
        ]
      },
...
      "WorkflowTemplateWithActiveVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b0000000-0000-4000-8000-000000000001"
          },
          "name": {
            "type": "string",
            "example": "Типовой цикл внедрения ИТ-продукта"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Базовый CLM-шаблон для вузов"
          },
          "activeVersion": {
            "nullable": true,
            "description": "Активная версия шаблона; null — если у шаблона ещё нет ни одной версии",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkflowTemplateVersionDto"
              }
            ]
          },
          "versions": {
            "description": "Все версии шаблона, по возрастанию номера",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowTemplateVersionSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "versions"
        ]
      },
...
      "CreateWorkflowTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Типовой цикл внедрения ИТ-продукта"
          },
          "description": {
            "type": "string",
            "example": "Базовый CLM-шаблон для вузов"
          },
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStatusInputDto"
            }
          },
          "transitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowTransitionInputDto"
            }
          }
        },
        "required": [
          "name",
          "statuses",
          "transitions"
        ]
      },
      "UpdateWorkflowTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Типовой цикл внедрения ИТ-продукта (v3)"
          },
          "description": {
            "type": "string",
            "example": "Добавлен статус приёмки после внедрения"
          },
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStatusInputDto"
            }
          },
          "transitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowTransitionInputDto"
            }
          },
          "migrateInstances": {
            "type": "boolean",
            "example": false,
            "description": "true — перевести взаимодействия предыдущей активной версии на новую (статус ищется по sourceStatusId). По умолчанию false: существующие процессы остаются на своей версии"
          }
        },
        "required": [
          "statuses",
          "transitions"
        ]
      },
...
    }
  }
}